pyproject.toml
[tool.poetry]
name = "CUNY_course"
version = "0.0.0"
description = "Cuny_Course"
authors = ["Lasse Funder Andersen <lass@dr.dk>"]
maintainers = ["lass-dr"]
readme = "README.md"
homepage = "https://github.com/drdk/CUNY_course"
repository = "https://github.com/drdk/CUNY_course"
documentation = "https://drdk.github.io/CUNY_course/"
packages = [
    { include = "CUNY_course"},
]
classifiers = [
    "Development Status :: 2 - Development",
]

[tool.poetry.urls]
Changelog = "https://github.com/drdk/CUNY_course/releases"

[tool.poetry.dependencies]
python = "^3.9.0"
click = ">=8.0.1"
python-dotenv = "^1.0.1"
openai = "^2.16.0"
pydantic = "^2.0.0"
numpy = "^1.26.4"
sentence-transformers = "^2.7.0"
faiss-cpu = "^1.8.0"
transformers = "^4.41.2"
torch = "^2.3.1"
gradio = "^4.44.0"

[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^24.3.0"}
ipykernel = "^6.29.4"
jupyter = "^1.0.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"

# [tool.poetry.scripts]
# Uncomment and add entry point if needed
# CUNY_course = "CUNY_course.__main__:main"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
